requirements.txt
crewai>=0.11.0
streamlit>=1.31.0
python-dotenv>=1.0.0
langchain-openai>=0.0.5
pytest>=8.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0 